speech-to-text-service/requirements.txt
scipy
paho.mqtt
openai